REFACTOR: Moved messages and tests of vesting module to own file.
