+++ pkg/indexer/decode/handle/test/vesting_test.go
@@ -1,0 +1,159 @@
+package handle_test
+
+import (
+	"cosmossdk.io/math"
+	"github.com/cosmos/cosmos-sdk/types"
+	cosmosVestingTypes "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
+	"github.com/dipdup-io/celestia-indexer/internal/storage"
+	storageTypes "github.com/dipdup-io/celestia-indexer/internal/storage/types"
+	"github.com/dipdup-io/celestia-indexer/internal/test_suite"
+	"github.com/dipdup-io/celestia-indexer/pkg/indexer/decode"
+	"github.com/fatih/structs"
+	"github.com/shopspring/decimal"
+	"github.com/stretchr/testify/assert"
+	"testing"
+)
+
+// MsgCreateVestingAccount
+
+func createMsgCreateVestingAccount() types.Msg {
+	m := cosmosVestingTypes.MsgCreateVestingAccount{
+		FromAddress: "celestia1j33593mn9urzydakw06jdun8f37shlucmhr8p6",
+		ToAddress:   "celestia1vsvx8n7f8dh5udesqqhgrjutyun7zqrgehdq2l",
+		Amount: types.Coins{
+			types.Coin{
+				Denom:  "utia",
+				Amount: math.NewInt(1000),
+			},
+		},
+		EndTime: 0,
+		Delayed: false,
+	}
+
+	return &m
+}
+
+func TestDecodeMsg_SuccessOnMsgCreateVestingAccount(t *testing.T) {
+	m := createMsgCreateVestingAccount()
+	blob, now := testsuite.EmptyBlock()
+	position := 0
+
+	dm, err := decode.Message(m, blob.Height, blob.Block.Time, position, storageTypes.StatusSuccess)
+
+	addressesExpected := []storage.AddressWithType{
+		{
+			Type: storageTypes.MsgAddressTypeFromAddress,
+			Address: storage.Address{
+				Id:         0,
+				Height:     blob.Height,
+				LastHeight: blob.Height,
+				Address:    "celestia1j33593mn9urzydakw06jdun8f37shlucmhr8p6",
+				Hash:       []byte{0x94, 0x63, 0x42, 0xc7, 0x73, 0x2f, 0x6, 0x22, 0x37, 0xb6, 0x73, 0xf5, 0x26, 0xf2, 0x67, 0x4c, 0x7d, 0xb, 0xff, 0x98},
+				Balance: storage.Balance{
+					Id:    0,
+					Total: decimal.Zero,
+				},
+			},
+		},
+		{
+			Type: storageTypes.MsgAddressTypeToAddress,
+			Address: storage.Address{
+				Id:         0,
+				Height:     blob.Height,
+				LastHeight: blob.Height,
+				Address:    "celestia1vsvx8n7f8dh5udesqqhgrjutyun7zqrgehdq2l",
+				Hash:       []byte{0x64, 0x18, 0x63, 0xcf, 0xc9, 0x3b, 0x6f, 0x4e, 0x37, 0x30, 0x0, 0x2e, 0x81, 0xcb, 0x8b, 0x27, 0x27, 0xe1, 0x0, 0x68},
+				Balance: storage.Balance{
+					Id:    0,
+					Total: decimal.Zero,
+				},
+			},
+		},
+	}
+
+	msgExpected := storage.Message{
+		Id:        0,
+		Height:    blob.Height,
+		Time:      now,
+		Position:  0,
+		Type:      storageTypes.MsgCreateVestingAccount,
+		TxId:      0,
+		Data:      structs.Map(m),
+		Namespace: nil,
+		Addresses: addressesExpected,
+	}
+
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(0), dm.BlobsSize)
+	assert.Equal(t, msgExpected, dm.Msg)
+	assert.Equal(t, addressesExpected, dm.Addresses)
+}
+
+// MsgCreatePeriodicVestingAccount
+
+func createMsgCreatePeriodicVestingAccount() types.Msg {
+	m := cosmosVestingTypes.MsgCreatePeriodicVestingAccount{
+		FromAddress:    "celestia1j33593mn9urzydakw06jdun8f37shlucmhr8p6",
+		ToAddress:      "celestia1vsvx8n7f8dh5udesqqhgrjutyun7zqrgehdq2l",
+		StartTime:      0,
+		VestingPeriods: nil,
+	}
+
+	return &m
+}
+
+func TestDecodeMsg_SuccessOnMsgCreatePeriodicVestingAccount(t *testing.T) {
+	msgCreatePeriodicVestingAccount := createMsgCreatePeriodicVestingAccount()
+	blob, now := testsuite.EmptyBlock()
+	position := 0
+
+	dm, err := decode.Message(msgCreatePeriodicVestingAccount, blob.Height, blob.Block.Time, position, storageTypes.StatusSuccess)
+
+	addressesExpected := []storage.AddressWithType{
+		{
+			Type: storageTypes.MsgAddressTypeFromAddress,
+			Address: storage.Address{
+				Id:         0,
+				Height:     blob.Height,
+				LastHeight: blob.Height,
+				Address:    "celestia1j33593mn9urzydakw06jdun8f37shlucmhr8p6",
+				Hash:       []byte{0x94, 0x63, 0x42, 0xc7, 0x73, 0x2f, 0x6, 0x22, 0x37, 0xb6, 0x73, 0xf5, 0x26, 0xf2, 0x67, 0x4c, 0x7d, 0xb, 0xff, 0x98},
+				Balance: storage.Balance{
+					Id:    0,
+					Total: decimal.Zero,
+				},
+			},
+		},
+		{
+			Type: storageTypes.MsgAddressTypeToAddress,
+			Address: storage.Address{
+				Id:         0,
+				Height:     blob.Height,
+				LastHeight: blob.Height,
+				Hash:       []byte{0x64, 0x18, 0x63, 0xcf, 0xc9, 0x3b, 0x6f, 0x4e, 0x37, 0x30, 0x0, 0x2e, 0x81, 0xcb, 0x8b, 0x27, 0x27, 0xe1, 0x0, 0x68},
+				Address:    "celestia1vsvx8n7f8dh5udesqqhgrjutyun7zqrgehdq2l",
+				Balance: storage.Balance{
+					Id:    0,
+					Total: decimal.Zero,
+				},
+			},
+		},
+	}
+
+	msgExpected := storage.Message{
+		Id:        0,
+		Height:    blob.Height,
+		Time:      now,
+		Position:  0,
+		Type:      storageTypes.MsgCreatePeriodicVestingAccount,
+		TxId:      0,
+		Data:      structs.Map(msgCreatePeriodicVestingAccount),
+		Namespace: nil,
+		Addresses: addressesExpected,
+	}
+
+	assert.NoError(t, err)
+	assert.Equal(t, uint64(0), dm.BlobsSize)
+	assert.Equal(t, msgExpected, dm.Msg)
+	assert.Equal(t, addressesExpected, dm.Addresses)
+}
