+++ pkg/indexer/decode/handle/vesting.go
@@ -1,0 +1,26 @@
+package handle
+
+import (
+	cosmosVestingTypes "github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
+	"github.com/dipdup-io/celestia-indexer/internal/storage"
+	storageTypes "github.com/dipdup-io/celestia-indexer/internal/storage/types"
+	"github.com/dipdup-io/celestia-indexer/pkg/types"
+)
+
+func MsgCreateVestingAccount(level types.Level, m *cosmosVestingTypes.MsgCreateVestingAccount) (storageTypes.MsgType, []storage.AddressWithType, error) {
+	msgType := storageTypes.MsgCreateVestingAccount
+	addresses, err := createAddresses(addressesData{
+		{t: storageTypes.MsgAddressTypeFromAddress, address: m.FromAddress},
+		{t: storageTypes.MsgAddressTypeToAddress, address: m.ToAddress},
+	}, level)
+	return msgType, addresses, err
+}
+
+func MsgCreatePeriodicVestingAccount(level types.Level, m *cosmosVestingTypes.MsgCreatePeriodicVestingAccount) (storageTypes.MsgType, []storage.AddressWithType, error) {
+	msgType := storageTypes.MsgCreatePeriodicVestingAccount
+	addresses, err := createAddresses(addressesData{
+		{t: storageTypes.MsgAddressTypeFromAddress, address: m.FromAddress},
+		{t: storageTypes.MsgAddressTypeToAddress, address: m.ToAddress},
+	}, level)
+	return msgType, addresses, err
+}
